requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
pyyaml>=6.0
psutil>=5.9.0  # Optional but recommended for accurate memory measurement
